move filesysten evaluate logic into base TransitionTool class this allows to switch between filesystem/stream t8n interactions for the derived classes
