+++ src/evm_transition_tool/transition_tool.py
@@ -35,6 +35,14 @@ class TransitionToolNotFoundInPath(Exception):
         super().__init__(message)
 
 
+def write_json_file(data: Dict[str, Any], file_path: str) -> None:
+    """
+    Write a JSON file to the given path.
+    """
+    with open(file_path, "w") as f:
+        json.dump(data, f, ensure_ascii=False, indent=4)
+
+
 def dump_files_to_directory(output_path: str, files: Dict[str, Any]) -> None:
     """
     Dump the files to the given directory.
@@ -106,6 +114,7 @@ class TransitionTool:
     statetest_subcommand: Optional[str] = None
     blocktest_subcommand: Optional[str] = None
     cached_version: Optional[str] = None
+    t8n_use_stream: bool = True
 
     # Abstract methods that each tool must implement
 
@@ -205,7 +214,9 @@ class TransitionTool:
             for subclass in subclasses:
                 if subclass.detect_binary(binary_output):
                     return subclass(binary=binary, **kwargs)
-        return cls.get_default_tool()(binary=binary, **kwargs)
+        default = cls.get_default_tool()(binary=binary, **kwargs)
+        default.t8n_use_stream = False
+        return default
 
     @classmethod
     def detect_binary(cls, binary_output: str) -> bool:
@@ -290,7 +301,7 @@ class TransitionTool:
                 traces.append(tx_traces)
         self.append_traces(traces)
 
-    def evaluate(
+    def _evaluateFilesystem(
         self,
         *,
         alloc: Any,
@@ -303,11 +314,131 @@ class TransitionTool:
         debug_output_path: str = "",
     ) -> Tuple[Dict[str, Any], Dict[str, Any]]:
         """
-        Executes `evm t8n` with the specified arguments.
-
-        If a client's `t8n` tool varies from the default behavior, this method
-        should be overridden.
+        Executes `evmone-t8n` with the specified arguments.
         """
+        if eips is not None:
+            fork_name = "+".join([fork_name] + [str(eip) for eip in eips])
+
+        temp_dir = tempfile.TemporaryDirectory()
+        os.mkdir(os.path.join(temp_dir.name, "input"))
+        os.mkdir(os.path.join(temp_dir.name, "output"))
+
+        input_contents = {
+            "alloc": alloc,
+            "env": env,
+            "txs": txs,
+        }
+
+        input_paths = {
+            k: os.path.join(temp_dir.name, "input", f"{k}.json") for k in input_contents.keys()
+        }
+        for key, file_path in input_paths.items():
+            write_json_file(input_contents[key], file_path)
+
+        output_paths = {
+            output: os.path.join("output", f"{output}.json") for output in ["alloc", "result"]
+        }
+        output_paths["body"] = os.path.join("output", "txs.rlp")
+
+        # Construct args for evmone-t8n binary
+        args = [
+            str(self.binary),
+            "--state.fork",
+            fork_name,
+            "--input.alloc",
+            input_paths["alloc"],
+            "--input.env",
+            input_paths["env"],
+            "--input.txs",
+            input_paths["txs"],
+            "--output.basedir",
+            temp_dir.name,
+            "--output.result",
+            output_paths["result"],
+            "--output.alloc",
+            output_paths["alloc"],
+            "--output.body",
+            output_paths["body"],
+            "--state.reward",
+            str(reward),
+            "--state.chainid",
+            str(chain_id),
+        ]
+
+        if self.trace:
+            args.append("--trace")
+
+        result = subprocess.run(
+            args,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
+
+        if debug_output_path:
+            if os.path.exists(debug_output_path):
+                shutil.rmtree(debug_output_path)
+            shutil.copytree(temp_dir.name, debug_output_path)
+            t8n_output_base_dir = os.path.join(debug_output_path, "t8n.sh.out")
+            t8n_call = " ".join(args)
+            for file_path in input_paths.values():  # update input paths
+                t8n_call = t8n_call.replace(
+                    os.path.dirname(file_path), os.path.join(debug_output_path, "input")
+                )
+            t8n_call = t8n_call.replace(  # use a new output path for basedir and outputs
+                temp_dir.name,
+                t8n_output_base_dir,
+            )
+            t8n_script = textwrap.dedent(
+                f"""\
+                #!/bin/bash
+                rm -rf {debug_output_path}/t8n.sh.out  # hard-coded to avoid surprises
+                mkdir -p {debug_output_path}/t8n.sh.out/output
+                {t8n_call}
+                """
+            )
+            dump_files_to_directory(
+                debug_output_path,
+                {
+                    "args.py": args,
+                    "returncode.txt": result.returncode,
+                    "stdout.txt": result.stdout.decode(),
+                    "stderr.txt": result.stderr.decode(),
+                    "t8n.sh+x": t8n_script,
+                },
+            )
+
+        if result.returncode != 0:
+            raise Exception("failed to evaluate: " + result.stderr.decode())
+
+        for key, file_path in output_paths.items():
+            output_paths[key] = os.path.join(temp_dir.name, file_path)
+
+        output_contents = {}
+        for key, file_path in output_paths.items():
+            if "txs.rlp" in file_path:
+                continue
+            with open(file_path, "r+") as file:
+                output_contents[key] = json.load(file)
+
+        if self.trace:
+            self.collect_traces(output_contents["result"]["receipts"], temp_dir, debug_output_path)
+
+        temp_dir.cleanup()
+
+        return output_contents["alloc"], output_contents["result"]
+
+    def _evaluateStream(
+        self,
+        *,
+        alloc: Any,
+        txs: Any,
+        env: Any,
+        fork_name: str,
+        chain_id: int = 1,
+        reward: int = 0,
+        eips: Optional[List[int]] = None,
+        debug_output_path: str = "",
+    ) -> Tuple[Dict[str, Any], Dict[str, Any]]:
         if eips is not None:
             fork_name = "+".join([fork_name] + [str(eip) for eip in eips])
 
@@ -400,6 +531,47 @@ class TransitionTool:
             temp_dir.cleanup()
 
         return output["alloc"], output["result"]
+
+    def evaluate(
+        self,
+        *,
+        alloc: Any,
+        txs: Any,
+        env: Any,
+        fork_name: str,
+        chain_id: int = 1,
+        reward: int = 0,
+        eips: Optional[List[int]] = None,
+        debug_output_path: str = "",
+    ) -> Tuple[Dict[str, Any], Dict[str, Any]]:
+        """
+        Executes `evm t8n` with the specified arguments.
+
+        If a client's `t8n` tool varies from the default behavior, this method
+        should be overridden.
+        """
+        if self.t8n_use_stream:
+            return self._evaluateStream(
+                alloc=alloc,
+                txs=txs,
+                env=env,
+                fork_name=fork_name,
+                chain_id=chain_id,
+                reward=reward,
+                eips=eips,
+                debug_output_path=debug_output_path,
+            )
+        else:
+            return self._evaluateFilesystem(
+                alloc=alloc,
+                txs=txs,
+                env=env,
+                fork_name=fork_name,
+                chain_id=chain_id,
+                reward=reward,
+                eips=eips,
+                debug_output_path=debug_output_path,
+            )
 
     def calc_state_root(
         self, *, alloc: Any, fork: Fork, debug_output_path: str = ""
